Fix attached accounts timing issue (#493)
I saw issues where the connected mongo database didn't always persist for attached accounts. It's because there is a timing issue related to loading the persisted accounts.

I suppose there was always a bit of a race condition here. For example, if the user expanded the 'Attached Accounts' node before we load the persisted accounts, they would not see their accounts.

Also fix "language client is not ready"